RavenDB-14173 cleanup
